Poll for action results for 30s instead of 60s
There's some strange issue with polling for 60s or more which I don't
comprehend. Go will never return the response, so only one poll attempt
is made. The HTTP request is somehow repeated and in another 30s
the connection is closed by the server. There shouldn't by any timeouts
to cause this, but waiting for less than 60s seems to solve it.